Changed rf.add() and rf.phase() to allow for uint32_t durations
Instead of limiting number of ticks to 15-bits (32767), RFControl allows for tick size to be any 32-bit number.  If ticks > 32767, RFControl adds repeated LOW or HIGH phases as needed to match full duration.  This provides for much more flexibility in creating pulse trains that include very long-duration "spaces" between repeats.

+++ src/extras/RFControl.h
@@ -22,8 +22,8 @@ class RFControl {
     void start(uint8_t nCycles=1, uint8_t tickTime=1);                                // starts transmission of pulses from internal data structure, repeated for numCycles, where each tick in pulse is tickTime microseconds long    
     
     void clear();                                         // clears transmitter memory
-    void add(uint16_t onTime, uint16_t offTime);          // adds pulse of onTime ticks HIGH followed by offTime ticks LOW
-    void phase(uint16_t nTicks, uint8_t phase);           // adds either a HIGH phase or LOW phase lasting numTicks ticks
+    void add(uint32_t onTime, uint32_t offTime);          // adds pulse of onTime ticks HIGH followed by offTime ticks LOW
+    void phase(uint32_t nTicks, uint8_t phase);           // adds either a HIGH phase or LOW phase lasting numTicks ticks
 };
 
 // Helper macro for creating your own storage of uint32_t data array elements - used with first variation of start() above
